requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
plotly==5.17.0
pandas==2.1.4
python-multipart==0.0.6
asyncio==3.4.3
aiofiles==23.2.1 